requirements-flyte.txt
# Flyte and workflow dependencies
flytekit>=1.10.0
flytekitplugins-pytorch>=1.10.0
flytekit[pytorch]>=1.10.0

# Core ML dependencies from your existing project
torch>=2.1.0
lightning>=2.1.0
tensordict>=0.2.0
jaxtyping>=0.2.0
beartype>=0.16.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Logging and monitoring
tensorboard>=2.15.0
wandb>=0.16.0

# Configuration and utilities
tomli>=2.0.0
pyyaml>=6.0
click>=8.1.0

# Additional dependencies that might be needed
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# For containerization
Pillow>=10.0.0
requests>=2.31.0